Fix filebrowser /run permissions issue with Memory emptyDir
The filebrowser pod was crashlooping with s6-overlay error:
"/run belongs to uid 0 instead of 1000, has insecure and/or
unworkable permissions"

This occurs because emptyDir volumes are created by Kubernetes with
root ownership. Using emptyDir with medium: Memory creates a tmpfs
that respects fsGroup permissions, allowing the pod to run as UID 1000.

Error: /run belongs to uid 0 instead of 1000