Fix cilium-gateway namespace - remove targetNamespace override
The CiliumGatewayClassConfig resources must be in kube-system namespace
but targetNamespace: network was forcing them into network namespace.

This caused the GatewayClass to not be accepted because it couldn't find
the CiliumGatewayClassConfig it references.

Removing targetNamespace allows resources to use their specified namespaces:
- CiliumGatewayClassConfig -> kube-system
- Gateway -> network
- ExternalSecret -> network

This fixes the Gateway API not being programmed and external-dns not
being able to publish DNS records to Mikrotik RouterOS.